requirements.txt
asgiref==3.4.1
boto3==1.23.10
botocore==1.26.10
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.0.12
cryptography==37.0.2
dataclasses==0.8
defusedxml==0.7.1
dj-database-url==0.5.0
Django==3.2.13
django-allauth==0.51.0
django-cors-headers==3.10.1
django-crispy-forms==1.13.0
django-filter==21.1
django-heroku==0.3.1
django-rest-knox==4.2.0
django-storages==1.12.3
Django-Verify-Email==2.0.3
djangorestframework==3.13.1
djongo==1.3.6
gunicorn==20.1.0
idna==3.3
jmespath==0.10.0
oauthlib==3.2.0
Pillow==8.4.0
psycopg2==2.9.3
pycparser==2.21
PyJWT==2.4.0
pymongo==3.12.1
python-dateutil==2.8.2
python-decouple==3.6
python3-openid==3.2.0
pytz==2022.1
requests==2.27.1
requests-oauthlib==1.3.1
s3transfer==0.5.2
six==1.16.0
social-auth-app-django==5.0.0
social-auth-core==4.3.0
sqlparse==0.2.4
typing_extensions==4.1.1
urllib3==1.26.9
whitenoise==5.3.0
